+++ src/quatrex/core/contact.py
@@ -1,3 +1,5 @@
+# Copyright (c) 2025-2026 ETH Zurich and the authors of the quatrex package.
+
 import itertools
 
 import numpy as np
@@ -6,28 +8,8 @@ from mpi4py.MPI import COMM_WORLD as comm
 from qttools import NDArray, obc, sparse, xp
 from qttools.nevp import NEVP, Beyn, Full
 from quatrex.core.compute_config import NEVPConfig
+from quatrex.core.kpoints import monkhorst_pack
 from quatrex.core.quatrex_config import OBCConfig
-
-
-# TODO move to some kid of utils
-def monkhorst_pack(size: tuple[int]) -> NDArray:
-    """Constructs a Monkhorst-Pack grid of k-points.
-
-    Parameters
-    ----------
-    size : tuple[int]
-        Grid dimensions as (nx, ny, nz) specifying the number of
-        k-points along each reciprocal lattice direction.
-
-    Returns
-    -------
-    kpts : NDArray
-        Array of k-points with shape (nx*ny*nz, 3). Each row contains
-        the (kx, ky, kz) coordinates of a k-point in reduced units.
-
-    """
-    kpts = np.indices(size).transpose((1, 2, 3, 0)).reshape((-1, 3))
-    return (kpts + 0.5) / size - 0.5
 
 
 class Contact:
